# utils/response.go
package utils

import "github.com/gin-gonic/gin"

type ApiResponse struct {
	Success    bool        `json:"success"`
	StatusCode int         `json:"status_code"`
	Message    string      `json:"message"`
	Data       interface{} `json:"data"`
}

func Success(ctx *gin.Context, statusCode int, message string, data interface{}) {
	ctx.JSON(statusCode, ApiResponse{
		Success:    true,
		StatusCode: statusCode,
		Message:    message,
		Data:       data,
	})
}

func Error(ctx *gin.Context, statusCode int, message string) {
	ctx.JSON(statusCode, ApiResponse{
		Success:    false,
		StatusCode: statusCode,
		Message:    message,
		Data:       nil,
	})
}